coursework_on_web/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-crontab==0.7.1
Pillow==9.4.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.1
sqlparse==0.4.3
